@@ -15,13 +15,16 @@
         "production": "mix --production"
     },
     "devDependencies": {
+        "@pinia/testing": "^0.1.3",
         "@types/jest": "^29.5.5",
         "@types/lodash": "^4.14.199",
         "@typescript-eslint/eslint-plugin": "^6.7.5",
         "@typescript-eslint/parser": "^6.7.5",
+        "@vue/compiler-sfc": "^3.3.8",
         "@vue/eslint-config-typescript": "^12.0.0",
-        "@vue/test-utils": "^1.3.6",
-        "@vue/vue2-jest": "^29.2.6",
+        "@vue/test-utils": "^2.0.0",
+        "@vue/vue3-jest": "^29.2.4",
+        "@vue/runtime-dom": "^3.3.8",
         "@wikimedia/codex": "^0.20.0",
         "@wikimedia/codex-design-tokens": "^0.20.0",
         "@wikimedia/codex-icons": "^0.20.0",
@@ -42,25 +45,19 @@
         "ts-jest": "^29.1.0",
         "ts-loader": "^9.5.0",
         "typescript": "^5.2.2",
-        "vue-loader": "^15.10.2",
-        "vue-template-compiler": "^2.6.14"
+        "vue-loader": "^16.0.0"
     },
     "dependencies": {
         "@inertiajs/inertia": "^0.11.0",
-        "@inertiajs/inertia-vue": "^0.8.0",
-        "@vue/composition-api": "^1.7.2",
+        "@inertiajs/inertia-vue3": "^0.6.0",
+        "@vue/compat": "^3.3.8",
         "@wmde/wikit-tokens": "^2.1.0-alpha.15",
         "@wmde/wikit-vue-components": "^2.1.0-alpha.16",
         "date-fns": "^2.30.0",
         "lodash": "^4.17.21",
+        "pinia": "^2.1.7",
         "ress": "^5.0.2",
-        "vue": "^2.6.14",
-        "vue-banana-i18n": "1.5.0",
-        "vuex": "^3.6.2"
-    },
-    "overrides": {
-        "@wikimedia/codex": {
-            "vue": "^2.6.14"
-        }
+        "vue": "^3.3.8",
+        "vue-banana-i18n": "2.3.0"
     }
 }
