@@ -68,14 +68,14 @@
 
 <script lang="ts">
 import { PropType } from 'vue';
-import { Link as InertiaLink } from '@inertiajs/inertia-vue';
+import { Link as InertiaLink } from '@inertiajs/inertia-vue3';
 import { Link as WikitLink } from '@wmde/wikit-vue-components';
 import { Button as LanguageSelectorButton, Icon } from '@wmde/wikit-vue-components';
 import AuthWidget from '../Components/AuthWidget.vue';
 import LanguageSelector from '../Components/LanguageSelector.vue';
 import WikidataToolFooter from '../Components/WikidataToolFooter.vue';
-import { DirectiveBinding } from 'vue/types/options';
-import defineComponent from '../types/defineComponent';
+import { DirectiveBinding, ComponentPublicInstance } from 'vue';
+import { defineComponent } from 'vue';
 import User from '../types/User';
 import languagedata from '@wikimedia/language-data';
 
@@ -99,7 +99,7 @@ export default defineComponent({
     },
     directives: {
         detectClickOutside: {
-            inserted(element: HTMLElement, binding: DirectiveBinding): void {
+          mounted(element: HTMLElement, binding: DirectiveBinding): void {
                 handleOutsideClick = (event: MouseEvent | TouchEvent): void => {
                     const callback = binding.value;
                     if (!element.contains(event.target as Node)) {
@@ -110,7 +110,7 @@ export default defineComponent({
                 document.addEventListener('click', handleOutsideClick);
                 document.addEventListener('touchstart', handleOutsideClick);
             },
-            unbind(): void {
+            unmounted(): void {
                 document.removeEventListener('click', handleOutsideClick);
                 document.removeEventListener('touchstart', handleOutsideClick);
             },
@@ -160,18 +160,22 @@ export default defineComponent({
             const headerTop = (this.$refs.header as HTMLElement).getBoundingClientRect().top;
             const userSectionTop = (this.$refs.userSection as HTMLElement).getBoundingClientRect().top;
             if( userSectionTop > headerTop ){
-                ((this.$refs.languageSelector as Vue).$el as HTMLElement).style.insetInlineEnd = 'unset';
-                ((this.$refs.languageSelector as Vue).$el as HTMLElement).style.insetInlineStart = '0';
+                ((this.$refs.languageSelector as ComponentPublicInstance)
+                    .$el as HTMLElement).style.insetInlineEnd = 'unset';
+                ((this.$refs.languageSelector as ComponentPublicInstance)
+                    .$el as HTMLElement).style.insetInlineStart = '0';
             } else {
-                ((this.$refs.languageSelector as Vue).$el as HTMLElement).style.insetInlineEnd = '0';
-                ((this.$refs.languageSelector as Vue).$el as HTMLElement).style.insetInlineStart = 'unset';
+                ((this.$refs.languageSelector as ComponentPublicInstance)
+                    .$el as HTMLElement).style.insetInlineEnd = '0';
+                ((this.$refs.languageSelector as ComponentPublicInstance)
+                    .$el as HTMLElement).style.insetInlineStart = 'unset';
             }
         },
         onWindowResize(): void {
             this.changeLanguageSelectorMenuDirection();
         },
     },
-    beforeDestroy () {
+    beforeUnmount () {
         this.resizeObserver.unobserve(this.$refs.contentWrap as Element)
     },
 });
