@@ -1,29 +1,25 @@
 import './bootstrap';
-import Vue from 'vue';
-import i18n from 'vue-banana-i18n';
-import { createStore } from './store';
-import { createInertiaApp } from '@inertiajs/inertia-vue';
-
-import i18nMessages from './lib/i18n';
+import {createApp, h} from 'vue';
+import {createPinia} from 'pinia';
+import {createInertiaApp} from '@inertiajs/inertia-vue3';
+import getI18nMessages, { I18nMessages } from './lib/i18n';
+import {createI18n} from 'vue-banana-i18n'
 import bubble from './lib/bubble';
 import Error from './Pages/Error.vue';
 import Layout from './Pages/Layout.vue';
-
-Vue.use(bubble);
-
-// Retrieve i18n messages and setup the Vue instance to handle them.
-async function setupI18n(locale: string): Promise<void>{
-    const messages = await i18nMessages(locale);
-    Vue.use(i18n, { locale, messages });
-}
 
 // Only bootstrap inertia if setup is successful. Display generic error
 // component otherwise
 (async () => {
     try {
-        await setupI18n(document.documentElement.lang);
-        const store = createStore();
-
+        const locale = document.documentElement.lang;
+        const i18nMessages = await getI18nMessages(locale);
+        const pinia = createPinia();
+        const i18nPlugin = createI18n({
+            locale: locale,
+            messages: i18nMessages,
+            wikilinks: true
+        });
         createInertiaApp({
             resolve: name => {
                 const page = require(`./Pages/${name}`).default;
@@ -33,22 +29,22 @@ async function setupI18n(locale: string): Promise<void>{
                 return page
             },
             setup({ el, app, props, plugin }) {
-                Vue.use(plugin)
-                new Vue({
-                    render: h => h(app, props),
-                    store
-                }).$mount(el);
+                createApp({
+                    render: () => h(app, props)
+                })
+                    .use(bubble)
+                    .use(i18nPlugin)
+                    .use(pinia)
+                    .use(plugin)
+                    .mount(el)
             }
         });
     } catch (e) {
-        new Vue({
-            render: h => h(Error, {
-                props: {
-                    title: 'Oops!',
-                    description: 'Something unexpected happened, but we are working on it... please try to refresh, or come back later.'
-                }
+        createApp({
+            render: () => h(Error, {
+                title: 'Oops!',
+                description: 'Something unexpected happened, but we are working on it... please try to refresh, or come back later.'
             }),
-        }).$mount('#app');
+        }).mount('#app');
     }
 })();
-
