@@ -3,282 +3,15 @@
     <inertia-head title="Mismatch Finder - Component Playground" />
     <h2>Mismatch Finder - Playground!</h2>
     <p>Feel free to throw any component you want to try, in here</p>
-    <wikit-button @click.native="controlled = true">Controlled Dialog!</wikit-button>
-    <wikit-button @click.native="$refs.standalone.show()">Standalone Dialog!</wikit-button>
-    <wikit-button @click.native="$refs.indismissible.show()">Indissmissable Dialog!</wikit-button>
-    <wikit-button @click.native="$refs.long.show()">Very Long Dialog!</wikit-button>
-    <wikit-button @click.native="$refs.closes.show()">Button Closed Dialog!</wikit-button>
-    <p>
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Inventore in nostrum ad autem enim cupiditate, facilis
-        recusandae vel adipisci, est nisi, consequatur explicabo sit quisquam eos cumque unde! Quaerat aliquam quis
-        nostrum numquam? Minima, quos quas, minus quo impedit ducimus ullam eveniet eaque porro aut tempore deleniti
-        nobis consequatur assumenda a incidunt nam sed, saepe sit laborum qui quam rerum laboriosam magnam! Illo numquam
-        atque harum in iusto accusantium delectus recusandae eveniet unde ut inventore cum distinctio ducimus, error
-        laboriosam sint? Modi qui dolores fugit ex debitis aliquam eligendi fugiat quos dolorem voluptatem, soluta
-        recusandae alias. Exercitationem labore debitis delectus commodi animi quod, nulla expedita, accusamus ea
-        doloremque obcaecati mollitia! Incidunt natus consequatur repudiandae at unde id exercitationem autem. Tenetur
-        commodi perspiciatis, eligendi adipisci molestiae nesciunt unde dicta ullam saepe velit, fugit aspernatur beatae
-        quos amet illo accusamus repellendus? Aliquam tenetur esse voluptate quisquam eligendi vitae, ducimus
-        voluptatibus, odit tempora quas dolores natus rerum nesciunt mollitia impedit distinctio nihil qui. Perferendis
-        recusandae autem possimus, aperiam molestiae dolores dignissimos, in atque pariatur eius, magnam illo debitis
-        animi veniam quo neque perspiciatis temporibus numquam minus facere mollitia! Quasi reiciendis provident dolores
-        animi eum quidem minima rerum, facilis aliquid, voluptates, ad quibusdam porro magni. Fugit inventore quis
-        cumque dolore beatae corporis deleniti unde? Beatae, recusandae aliquid ipsum unde consequatur aut harum
-        voluptatibus doloribus expedita omnis rerum praesentium quaerat quisquam optio sed dolor enim autem et! Magnam
-        asperiores nam delectus nihil iusto voluptatibus ad et, esse sed sint sapiente cumque eos fugit vero! Repellat
-        doloremque quibusdam ullam dicta possimus temporibus aperiam deserunt enim est expedita. Deleniti in impedit
-        architecto itaque dolores vero reprehenderit ea reiciendis. Impedit esse, molestias explicabo possimus
-        temporibus consectetur vitae delectus similique sint facere. Magni magnam obcaecati error ducimus saepe maxime
-        velit at, soluta nihil tempora praesentium sit molestiae nemo consectetur, et earum voluptas? Reiciendis
-        provident corporis dicta. Ad, facilis officia omnis unde, corrupti itaque, quaerat similique nisi ab libero
-        fugiat facere. Nostrum sed eos, adipisci, eaque quibusdam molestiae rem id expedita, eveniet veritatis magni
-        maxime dolorem ipsa qui minus incidunt ipsum. Aliquam sequi, magni voluptates blanditiis aspernatur tempora id
-        iste explicabo repudiandae corporis voluptatibus! Ea voluptate eaque deleniti molestiae ullam, accusamus atque
-        quaerat facilis provident autem dolores ab rerum aliquid necessitatibus porro est quam sequi. Optio omnis quos
-        provident dolor, perferendis reiciendis consectetur in nulla eos autem ducimus iusto saepe asperiores amet
-        vitae, placeat eum, perspiciatis delectus harum. Itaque nesciunt libero, sunt provident corrupti animi eos
-        laudantium modi minus deleniti atque quibusdam natus, qui tenetur odio eum cum quam. Necessitatibus natus
-        aliquid totam, odio qui dolore quibusdam error quaerat odit illum ex veniam consectetur quasi laboriosam
-        voluptatibus deserunt dolores id, corrupti tempora veritatis! Autem a dicta aspernatur aliquam, quo similique
-        natus tempora officia fugiat facilis ducimus dolorum dolore mollitia accusantium magnam voluptatem inventore ex
-        odit voluptas numquam ab laboriosam unde? Nam eos ullam praesentium provident, veniam tempora libero aspernatur
-        qui, ducimus officiis sequi tempore? Reprehenderit ea aut consequatur animi asperiores quidem inventore neque
-        accusantium suscipit laboriosam accusamus magni molestias, unde quia nam quisquam blanditiis recusandae nisi
-        perferendis beatae explicabo in reiciendis atque. Consectetur provident animi atque nulla commodi assumenda,
-        illo odit id est, voluptas totam repellendus ut minus voluptates, et facere deleniti nemo delectus labore at
-        facilis eaque repudiandae sequi. Voluptatum consequatur deserunt delectus eaque tempore. Aperiam soluta quidem
-        pariatur ad numquam ea deserunt dignissimos, sunt fugit similique quo quos est itaque nihil! Dolore
-        exercitationem praesentium numquam distinctio veniam unde, ipsum expedita, nesciunt fuga corporis dolorum est
-        architecto. Architecto necessitatibus nisi velit facilis a ipsa, voluptatibus in quis illo, vel non rem?
-        Excepturi pariatur accusamus, animi odit ducimus ab ratione dicta? Dolore atque placeat, eius quisquam ea
-        officia suscipit labore qui in voluptatum cupiditate harum accusantium ullam illum magni minima commodi
-        dignissimos! Reiciendis, accusantium dolores, illo officiis molestias sapiente asperiores corporis impedit
-        delectus harum at quod, deserunt molestiae. Assumenda iste at corporis earum, autem fugiat obcaecati sapiente,
-        nihil ipsa veniam accusantium? Officiis delectus cum placeat corporis minus dicta perferendis, amet obcaecati
-        voluptates fugit vel excepturi rem, sunt tempora quod, quis natus nulla maxime voluptatum voluptatem vitae
-        exercitationem. Explicabo repellat animi odio incidunt itaque qui quaerat, fuga at dolorem reiciendis fugit quod
-        adipisci cumque est ut id ad distinctio blanditiis sint esse quidem doloribus maiores in? Voluptatibus
-        laboriosam ducimus non magni incidunt voluptates omnis illo quasi labore dolores veritatis pariatur vero,
-        eveniet dolorem praesentium deserunt a accusantium unde modi exercitationem natus fugiat placeat! Accusamus quae
-        illum voluptate et ipsam sit ullam voluptates. Saepe quo nesciunt distinctio debitis neque deleniti! Delectus,
-        sint? Dolorem pariatur quo at cum quaerat obcaecati adipisci reprehenderit facere vero? Culpa minima sapiente in
-        aut quasi, ipsum architecto tempora sequi, quis quaerat laudantium quae aliquid alias cupiditate officiis,
-        reprehenderit doloribus nemo ut labore? Doloremque animi ad itaque autem similique perspiciatis eos explicabo,
-        tempore esse expedita facere labore doloribus, beatae temporibus est necessitatibus quas? Minima nulla expedita
-        ab ratione harum eaque illum, voluptatum exercitationem magnam delectus est quod dicta voluptas impedit omnis
-        perferendis possimus blanditiis quibusdam, corporis fuga? Repudiandae, architecto mollitia id fugiat iusto sint
-        ullam voluptas nesciunt maxime hic vel illum maiores laborum quod ducimus ipsa a ad. Eligendi est expedita sunt.
-        Fugiat unde temporibus quam at deleniti sit excepturi molestias repellendus sed iste, officiis atque, delectus
-        est repudiandae tenetur illum sint perspiciatis doloribus corrupti possimus, enim culpa. Quas veritatis nulla
-        laudantium excepturi maiores tenetur temporibus beatae nisi culpa cupiditate rem provident, sit natus ea vitae
-        at amet in. Error eaque dolores optio, adipisci perspiciatis at dicta aperiam, a praesentium aut quaerat veniam
-        animi alias quam voluptate iusto, id enim necessitatibus minus recusandae soluta? Aliquid saepe fugit error
-        explicabo dolore voluptates minus fugiat voluptas eligendi excepturi. Nostrum sunt aliquam neque tempora
-        mollitia. Alias facere impedit natus maxime eligendi, voluptatem aut quidem magni dignissimos quaerat vel
-        voluptatum at sequi tempora soluta libero iste nulla, eaque odit rem debitis cumque. Nam laudantium delectus
-        officiis! Fugiat exercitationem, sed esse, quisquam aperiam perspiciatis vel accusantium veniam cupiditate,
-        eaque obcaecati doloremque reprehenderit cum dignissimos nihil! Fuga perferendis debitis aperiam exercitationem
-        facere est dolore distinctio?
-    </p>
-
-    <wikit-dialog class="fancy-dialog"
-        title="Controlled Dialog"
-        :actions="[
-            {
-                label: 'Test Primary',
-                namespace: 'primary'
-            },
-            {
-                label: 'Test Secondary',
-                namespace: 'secondary'
-            }
-        ]"
-        ref="controlledDialog"
-        :visible.sync="controlled"
-        dismiss-button>
-        <h2>I am a dialog!</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reiciendis provident impedit iste iure perferendis,
-        necessitatibus enim, laboriosam tempora odio sed voluptatum modi doloribus error soluta dignissimos obcaecati.
-        Ad laborum vero ipsa accusamus a? Doloribus, aliquid impedit, rem modi perspiciatis id omnis, quis nemo fuga
-        repellendus adipisci? Soluta molestiae ullam doloremque aspernatur delectus rerum aliquid. Culpa dolores dicta
-        qui, tenetur odio placeat facilis enim fugiat sunt doloribus cupiditate officia facere corporis unde? Recusandae
-        eaque, id molestias porro velit mollitia similique exercitationem totam rem. Nisi quisquam commodi velit tempore
-        pariatur, esse voluptate architecto dolor mollitia? Eligendi minus ipsum aliquam voluptate vitae eveniet.</p>
-        <wikit-checkbox label="With a fancy checkbox!" :checked="false" />
-    </wikit-dialog>
-
-    <wikit-dialog class="fancy-dialog"
-        title="Standalone Dialog"
-        :actions="[
-            {
-                label: 'Test Primary',
-                namespace: 'primary'
-            },
-            {
-                label: 'Test Secondary',
-                namespace: 'secondary'
-            }
-        ]"
-        ref="standalone"
-        :visible="standalone"
-        dismiss-button>
-        <h2>I am a dialog!</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reiciendis provident impedit iste iure perferendis,
-        necessitatibus enim, laboriosam tempora odio sed voluptatum modi doloribus error soluta dignissimos obcaecati.
-        Ad laborum vero ipsa accusamus a? Doloribus, aliquid impedit, rem modi perspiciatis id omnis, quis nemo fuga
-        repellendus adipisci? Soluta molestiae ullam doloremque aspernatur delectus rerum aliquid. Culpa dolores dicta
-        qui, tenetur odio placeat facilis enim fugiat sunt doloribus cupiditate officia facere corporis unde? Recusandae
-        eaque, id molestias porro velit mollitia similique exercitationem totam rem. Nisi quisquam commodi velit tempore
-        pariatur, esse voluptate architecto dolor mollitia? Eligendi minus ipsum aliquam voluptate vitae eveniet.</p>
-        <wikit-checkbox label="With a fancy checkbox!" :checked="false" />
-    </wikit-dialog>
-
-    <wikit-dialog class="fancy-dialog"
-        title="Indismissible Dialog"
-        :actions="[
-            {
-                label: 'Test Primary',
-                namespace: 'primary'
-            },
-            {
-                label: 'Test Secondary',
-                namespace: 'secondary'
-            }
-        ]"
-        ref="indismissible">
-        <h2>I am a dialog!</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reiciendis provident impedit iste iure perferendis,
-        necessitatibus enim, laboriosam tempora odio sed voluptatum modi doloribus error soluta dignissimos obcaecati.
-        Ad laborum vero ipsa accusamus a? Doloribus, aliquid impedit, rem modi perspiciatis id omnis, quis nemo fuga
-        repellendus adipisci? Soluta molestiae ullam doloremque aspernatur delectus rerum aliquid. Culpa dolores dicta
-        qui, tenetur odio placeat facilis enim fugiat sunt doloribus cupiditate officia facere corporis unde? Recusandae
-        eaque, id molestias porro velit mollitia similique exercitationem totam rem. Nisi quisquam commodi velit tempore
-        pariatur, esse voluptate architecto dolor mollitia? Eligendi minus ipsum aliquam voluptate vitae eveniet.</p>
-        <wikit-checkbox label="With a fancy checkbox!" :checked="false" />
-    </wikit-dialog>
-
-    <wikit-dialog class="fancy-dialog"
-        title="Very Long Dialog"
-        :actions="[
-            {
-                label: 'Test Primary',
-                namespace: 'primary'
-            }
-        ]"
-        ref="long">
-        <h2>I am a dialog!</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reiciendis provident impedit iste iure perferendis,
-        necessitatibus enim, laboriosam tempora odio sed voluptatum modi doloribus error soluta dignissimos obcaecati.
-        Ad laborum vero ipsa accusamus a? Doloribus, aliquid impedit, rem modi perspiciatis id omnis, quis nemo fuga
-        repellendus adipisci? Soluta molestiae ullam doloremque aspernatur delectus rerum aliquid. Culpa dolores dicta
-        qui, tenetur odio placeat facilis enim fugiat sunt doloribus cupiditate officia facere corporis unde? Recusandae
-        eaque, id molestias porro velit mollitia similique exercitationem totam rem. Nisi quisquam commodi velit tempore
-        pariatur, esse voluptate architecto dolor mollitia? Eligendi minus ipsum aliquam voluptate vitae eveniet.</p>
-        <p>Lorem ipsum dolor sit, amet consectetur adipisicing elit. Nesciunt excepturi iure eius deleniti quisquam
-        soluta ipsam beatae, incidunt consectetur expedita tenetur neque fuga ab necessitatibus id facilis nulla ut
-        dolore debitis dolor sequi corrupti quae! Enim repellat totam impedit quia iste assumenda minima tempora soluta,
-        laudantium molestias harum voluptas. Delectus nobis voluptatibus optio porro rem similique unde veniam
-        aspernatur ratione fugiat neque ab ipsum distinctio, corporis provident aperiam dolore accusantium consequatur,
-        cumque sed illum deserunt.</p>
-        <p>Dolor, atque, natus beatae, quam voluptate nesciunt laboriosam labore impedit accusantium explicabo minima
-        modi. Sequi eum accusantium nesciunt rem cum quis, totam quidem dolorum facere sunt aliquid perspiciatis eius
-        nemo doloribus quas ab quod libero repudiandae inventore minima fugit quaerat vero. Eaque nesciunt dolore
-        tempore nam exercitationem, necessitatibus amet magni maiores consequatur assumenda tenetur hic sunt at
-        doloremque. Amet incidunt veritatis distinctio officiis magni reprehenderit nostrum minima quod voluptatibus
-        consequuntur, voluptatem, soluta ipsam impedit quam?</p>
-        <p>Architecto, voluptates necessitatibus voluptate, nobis dicta culpa nesciunt porro temporibus tenetur sequi
-        deserunt sint quis nostrum laudantium optio, laboriosam possimus. Maiores tempora magni similique! Corrupti
-        cumque porro ipsum earum cum mollitia enim natus non, rem possimus saepe necessitatibus quo neque consectetur
-        quae alias maiores impedit, eos fugit officiis commodi ab tempore tempora consequatur! Laudantium magni
-        consequatur error quidem eveniet explicabo dolorem, sequi ut quis tempore earum quaerat vero asperiores omnis.
-        Impedit facere corporis architecto placeat!</p>
-        <p>Provident, impedit maiores enim quod exercitationem veritatis praesentium? Iure aperiam molestiae non
-        voluptas temporibus voluptatem soluta neque animi laboriosam itaque voluptatum eum possimus est similique quia
-        dolor adipisci perspiciatis et, quibusdam explicabo veritatis libero optio a doloribus! Aperiam laudantium ullam
-        ipsum fugiat iste culpa architecto, quaerat possimus id labore aut accusantium doloremque porro, harum commodi,
-        nemo natus sapiente quae dignissimos itaque sed error? Id ipsa ipsam natus sint sit consequuntur explicabo
-        impedit asperiores autem recusandae!</p>
-        <p>Doloribus, eaque? Odio harum hic libero vel, nisi nulla est deserunt laboriosam repudiandae cupiditate saepe
-        mollitia impedit, eveniet tempora consequatur repellendus asperiores, molestiae voluptatibus? Iure aspernatur
-        asperiores quibusdam. Quae totam aliquam eos exercitationem. Totam quo eaque at molestias dolorem voluptate ab
-        pariatur! Explicabo porro maiores sequi sapiente commodi nam placeat velit eaque quidem, exercitationem eum quas
-        doloremque obcaecati autem sint modi saepe cumque harum soluta possimus? Animi temporibus consectetur ex eaque
-        similique. Rem, minima quos.</p>
-        <p>Aliquam nobis temporibus magnam id culpa? Aspernatur beatae ullam quidem repellendus ab blanditiis modi quasi
-        fugiat soluta tempora a iusto quod illum dicta vitae debitis doloremque, qui eos repudiandae placeat sit.
-        Consequuntur, qui cum cumque neque sapiente quasi expedita esse amet commodi minus aliquid mollitia tempora
-        dolore rerum numquam, modi iure sunt aut. Sint accusamus delectus, animi saepe voluptatibus nostrum. Eveniet
-        voluptatum, consectetur dolore est, atque voluptate, doloribus ratione aperiam eligendi ad expedita voluptatibus
-        corporis?</p>
-        <p>Sequi illum inventore enim deleniti eius magnam, totam ratione. Sint architecto facilis quas ducimus ea
-        obcaecati molestias iure, inventore pariatur, molestiae at amet vero provident asperiores facere maiores
-        repellendus. Repellendus consequuntur sed magnam mollitia, labore voluptate odio vel ducimus cum necessitatibus
-        saepe, dolores commodi temporibus sunt provident in praesentium veniam maxime? Deserunt voluptatem asperiores
-        dolores tenetur porro illo perferendis soluta doloremque beatae ea! Ipsam iure corrupti ad, excepturi voluptas
-        animi modi alias minima? Pariatur, quaerat!</p>
-        <p>Amet error facere vel eveniet officiis necessitatibus omnis ipsum minima, quia voluptatum asperiores vero
-        nesciunt libero distinctio rerum excepturi incidunt alias quae, similique possimus, cupiditate dicta. Numquam at
-        libero non dicta asperiores corporis in, eaque similique unde, ducimus animi nesciunt! Laborum id voluptate
-        porro atque nobis perspiciatis, mollitia, quia labore repudiandae, eligendi delectus eos debitis amet expedita
-        dolore vitae accusamus asperiores. Aperiam ipsam animi optio ratione nostrum non incidunt assumenda voluptates?
-        Laboriosam vel atque nisi?</p>
-        <p>Molestiae asperiores autem consectetur distinctio eum nulla mollitia cum, tempore quae velit rerum ut. Nam,
-        illum quas. Unde necessitatibus repellendus magni voluptates, natus id praesentium ad labore eveniet, ducimus,
-        iste quisquam assumenda porro est aliquam. Sit reprehenderit aspernatur explicabo est totam. Autem, soluta.
-        Molestias, vero quam unde voluptatum inventore quibusdam libero consectetur officia illum excepturi, facilis,
-        quas facere minus cum deleniti commodi accusantium quae iste maxime. Repellendus, reiciendis! Architecto velit
-        rem labore. Repellendus, distinctio labore.</p>
-        <p>Architecto animi nisi natus minima ullam, eligendi, in ex, dolore eveniet dolorum quisquam soluta. Rerum
-        cupiditate debitis autem, laboriosam maiores facilis illum voluptatem quasi excepturi repudiandae sequi.
-        Perferendis quasi quae, unde dolores similique molestiae expedita enim. Saepe possimus excepturi at libero
-        facilis iure eaque deserunt recusandae, quam reprehenderit repudiandae voluptatibus id non, autem accusantium
-        sint expedita. Incidunt tempore vel vitae debitis, nostrum quas accusamus. Consequatur aspernatur praesentium
-        neque aliquam recusandae velit repellat perferendis unde explicabo.</p>
-    </wikit-dialog>
-
-    <wikit-dialog class="fancy-dialog"
-        title="Button Closed Dialog"
-        :actions="[
-            {
-                label: 'Boopy',
-                namespace: 'primary-boop'
-            },
-            {
-                label: 'Schmoopy',
-                namespace: 'secondary-schmoop'
-            }
-        ]"
-        @action="(namespace, dialog) => namespace === 'primary-boop' ? dialog.hide() : ''"
-        ref="closes">
-        <h2>I am a dialog!</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Reiciendis provident impedit iste iure perferendis,
-        necessitatibus enim, laboriosam tempora odio sed voluptatum modi doloribus error soluta dignissimos obcaecati.
-        Ad laborum vero ipsa accusamus a? Doloribus, aliquid impedit, rem modi perspiciatis id omnis, quis nemo fuga
-        repellendus adipisci? Soluta molestiae ullam doloremque aspernatur delectus rerum aliquid. Culpa dolores dicta
-        qui, tenetur odio placeat facilis enim fugiat sunt doloribus cupiditate officia facere corporis unde? Recusandae
-        eaque, id molestias porro velit mollitia similique exercitationem totam rem. Nisi quisquam commodi velit tempore
-        pariatur, esse voluptate architecto dolor mollitia? Eligendi minus ipsum aliquam voluptate vitae eveniet.</p>
-    </wikit-dialog>
   </div>
 </template>
 
 <script lang="ts">
-    import { Head as InertiaHead } from '@inertiajs/inertia-vue'
-    import defineComponent from '../types/defineComponent';
-
-    import {
-        Button as WikitButton,
-        Checkbox as WikitCheckbox,
-        Dialog as WikitDialog
-    } from '@wmde/wikit-vue-components';
+    import { defineComponent } from 'vue';
+    import { Head as InertiaHead } from '@inertiajs/inertia-vue3'
 
     export default defineComponent({
-        components: {
-            InertiaHead,
-            WikitButton,
-            WikitCheckbox,
-            WikitDialog
-        },
+        components: { InertiaHead },
         props: {
             user: Object,
         },
