@@ -15,7 +15,7 @@
                     id="instructions-button"
                     variant="quiet"
                     type="progressive"
-                    @click.native="showInstructionsDialog"
+                    @click.native="instructionsDialog = true"
                 >
                     <template #prefix>
                         <icon type="info-outlined" size="medium" color="inherit"/>
@@ -24,15 +24,16 @@
                 </wikit-button>
             </header>
 
-            <wikit-dialog id="instructions-dialog"
+            <cdx-dialog id="instructions-dialog"
                 :title="$i18n('instructions-dialog-title')"
-                ref="inctructionsDialog"
-                :actions="[{
+                v-model:open="instructionsDialog"
+                :primary-action="{
                     label: $i18n('confirm-dialog-button'),
-                    namespace: 'instructions-confirm'
-                }]"
-                @action="(_, dialog) => dialog.hide()"
-                dismiss-button
+                    namespace: 'instructions-confirm',
+                    actionType: 'progressive'
+                }"
+                @primary="() => instructionsDialog = false"
+                close-button-label="X"
             >
                 <p>{{ $i18n('instructions-dialog-message-upload-info-description') }}</p>
                 <p>{{ $i18n('instructions-dialog-message-intro') }}</p>
@@ -44,7 +45,7 @@
                     <li>{{ $i18n('instructions-dialog-message-instruction-none') }}</li>
                     <li>{{ $i18n('instructions-dialog-message-instruction-pending') }}</li>
                 </ul>
-            </wikit-dialog>
+            </cdx-dialog>
             <p id="about-description" >
                 {{ $i18n('results-page-description') }}
             </p>
@@ -57,7 +58,7 @@
                 <span>{{ $i18n('no-mismatches-found-message') }}</span>
                 <span class="message-link" v-for="item_id in notFoundItemIds" :key="item_id">
                     <wikit-link
-                        :href="`https://www.wikidata.org/wiki/${item_id}`" target="_blank">
+                        :href="`https://www.wikidata.org/wiki/${String(item_id)}`" target="_blank">
                         {{labels[item_id]}} ({{item_id}})
                     </wikit-link>
                 </span>
@@ -73,10 +74,10 @@
         <section id="results" v-if="Object.keys(results).length">
             <section class="item-mismatches"
                 v-for="(mismatches, item, idx) in results"
-                :id="`item-mismatches-${item}`"
+                :id="`item-mismatches-${String(item)}`"
                 :key="idx">
                 <h2 class="h4">
-                    <wikit-link :href="`https://www.wikidata.org/wiki/${item}`" target="_blank">
+                    <wikit-link :href="`https://www.wikidata.org/wiki/${String(item)}`" target="_blank">
                         {{labels[item]}} ({{item}})
                     </wikit-link>
                 </h2>
@@ -89,7 +90,7 @@
                         <Message class="form-success-message" type="success" v-if="lastSubmitted === item">
                             <span>{{ $i18n('changes-submitted-message') }}</span>
                             <span class="message-link">
-                                <wikit-link :href="`https://www.wikidata.org/wiki/${item}`" target="_blank">
+                                <wikit-link :href="`https://www.wikidata.org/wiki/${String(item)}`" target="_blank">
                                     {{labels[item]}} ({{item}})
                                 </wikit-link>
                             </span>
@@ -108,16 +109,18 @@
                 </form>
             </section>
         </section>
-        <wikit-dialog class="confirmation-dialog"
+        <cdx-dialog id="results-confirmation-dialog"
             :title="$i18n('confirmation-dialog-title')"
-            ref="confirmation"
-            :actions="[{
+            v-model:open="confirmationDialog"
+            :primary-action="{
                 label: $i18n('confirmation-dialog-button'),
-                namespace: 'next-steps-confirm'
-            }]"
-            @action="_handleConfirmation"
-            @dismissed="disableConfirmation = false"
-            dismiss-button
+                namespace: 'next-steps-confirm',
+                actionType: 'progressive'
+
+            }"
+            @update:open="disableConfirmation = false"
+            @primary="_handleConfirmation"
+            close-button-label="X"
         >
             <p>{{ $i18n('confirmation-dialog-message-intro') }}</p>
             <ul>
@@ -129,28 +132,29 @@
                 :label="$i18n('confirmation-dialog-option-label')"
                 :checked.sync="disableConfirmation"
             />
-        </wikit-dialog>
+        </cdx-dialog>
     </div>
 </template>
 
 <script lang="ts">
     import { PropType } from 'vue';
-    import { mapMutations } from 'vuex';
+    import { useStore } from '../store';
     import isEmpty from 'lodash/isEmpty';
-    import { Head as InertiaHead } from '@inertiajs/inertia-vue';
+    import { Head as InertiaHead } from '@inertiajs/inertia-vue3';
     import {
         Link as WikitLink,
         Button as WikitButton,
         Checkbox,
-        Dialog as WikitDialog,
         Icon,
         Message } from '@wmde/wikit-vue-components';
+
+    import { CdxDialog } from "@wikimedia/codex";
 
     import LoadingOverlay from '../Components/LoadingOverlay.vue';
     import MismatchesTable from '../Components/MismatchesTable.vue';
     import Mismatch, {ReviewDecision, LabelledMismatch} from '../types/Mismatch';
     import User from '../types/User';
-    import defineComponent from '../types/defineComponent';
+    import { defineComponent } from 'vue';
     import axios from 'axios';
 
     interface MismatchDecision {
@@ -183,7 +187,9 @@
         disableConfirmation: boolean,
         pageDirection: string,
         requestError: boolean,
-        lastSubmitted: string
+        lastSubmitted: string,
+        instructionsDialog: boolean,
+        confirmationDialog: boolean
     }
 
     export default defineComponent({
@@ -194,9 +200,9 @@
             MismatchesTable,
             WikitLink,
             WikitButton,
-            WikitDialog,
             Checkbox,
-            Message
+            Message,
+            CdxDialog
         },
         props: {
             user: {
@@ -226,8 +232,9 @@
             },
         },
         mounted(){
-            if(!this.$store.state.lastSearchedIds) {
-                this.saveSearchedIds( this.item_ids.join('\n') );
+            const store = useStore();
+            if(!store.lastSearchedIds) {
+                store.saveSearchedIds( this.item_ids.join('\n') );
             }
 
             this.pageDirection = window.getComputedStyle(document.body).direction;
@@ -253,15 +260,12 @@
                 disableConfirmation: false,
                 pageDirection: 'ltr',
                 requestError: false,
-                lastSubmitted: ''
+                lastSubmitted: '',
+                instructionsDialog: false,
+                confirmationDialog: false
             }
         },
         methods: {
-            showInstructionsDialog() {
-                /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
-                const dialog = this.$refs.inctructionsDialog as any;
-                dialog.show();
-            },
             addLabels(mismatches: Mismatch[]): LabelledMismatch[]{
                 // The following callback maps existing mismatches to extended
                 // mismatch objects which include labels, by looking up any
@@ -307,7 +311,6 @@
                 // Defaulting to any, as the alternative presents us with
                 // convoluted and unnecessary syntax.
                 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                const confirmationDialog = this.$refs.confirmation as any;
                 // eslint-disable-next-line @typescript-eslint/no-explicit-any
                 const overlay = this.$refs.overlay as any;
 
@@ -323,7 +326,7 @@
                     this.showSubmitConfirmation(item);
 
                     if(!this.disableConfirmation){
-                        confirmationDialog.show();
+                      this.confirmationDialog = true;
                     }
                 } catch(e) {
                     this.requestError = true;
@@ -357,7 +360,7 @@
             // understand component instances and complains about usage of the
             // hide method otherwise.
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            _handleConfirmation(_ : string, dialog: any){
+            _handleConfirmation(){
                 const { disableConfirmation, user } = this;
 
                 // Do nothing if there is no user
@@ -370,9 +373,8 @@
                     window.localStorage.setItem(`mismatch-finder_user-settings_${user.id}`, storageData);
                 }
 
-                dialog.hide();
-            },
-            ...mapMutations(['saveSearchedIds'])
+              this.confirmationDialog = false;
+            }
         }
     });
 </script>
