@@ -40,7 +40,8 @@ class HomePage extends Page
     {
         return [
             '@form' => '#items-form',
-            '@items-input' => '@form textarea'
+            '@items-input' => '@form textarea',
+            '@items-input-validation-message' => '@form .wikit-TextArea .wikit-ValidationMessage'
         ];
     }
 }
