@@ -1,10 +1,10 @@
-import { mount, createLocalVue } from '@vue/test-utils';
-import Vuex from 'vuex'
-import Home, { MAX_NUM_IDS } from '@/Pages/Home.vue'
+import { mount } from '@vue/test-utils';
+import { createTestingPinia } from '@pinia/testing';
+import Home, { MAX_NUM_IDS } from '@/Pages/Home.vue';
 
 // Stub the inertia vue components module entirely so that we don't run into
 // issues with the Head component.
-jest.mock('@inertiajs/inertia-vue', () => ({}));
+jest.mock('@inertiajs/inertia-vue3', () => ({}));
 
 describe('Home.vue', () => {
 
@@ -15,19 +15,16 @@ describe('Home.vue', () => {
         },
     }
 
-    const localVue = createLocalVue();
-    localVue.use(Vuex);
 
     it('sanitises input with empty lines', async () => {
 
         const itemsInput = 'Q1\n\nQ2\n';
 
-        const store = new Vuex.Store({});
-
         const wrapper = mount(Home, {
-            mocks,
-            localVue,
-            store,
+            global: {
+                mocks,
+                plugins: [createTestingPinia()],
+            },
             data() {
                 return {
                     form: {
@@ -43,36 +40,49 @@ describe('Home.vue', () => {
     it('restores lastSearchIds value from store on page load', async () => {
 
         const itemsInput = 'Q1\n\nQ2\n';
-        const store = new Vuex.Store({state: {lastSearchedIds: itemsInput} });
 
         const wrapper = mount(Home, {
-            mocks,
-            localVue,
-            store
+            global: {
+                mocks,
+                plugins: [createTestingPinia({
+                    initialState: {
+                        storeId: {
+                            lastSearchedIds: itemsInput
+                        }
+                    }
+                })]
+            }
         });
 
         expect( wrapper.vm.form.itemsInput ).toEqual(itemsInput);
     });
 
     it('shows dialog after clicking the more info button', async () => {
-        const store = new Vuex.Store();
 
-        const wrapper = mount(Home, { mocks, localVue, store });
+        const wrapper = mount(Home, {
+            attachTo: document.body,
+            global:
+            {
+                mocks,
+                plugins: [createTestingPinia()],
+                stubs: {
+                    teleport: true,
+                    transition: true
+                }
+            }});
         await wrapper.find('#faq-button').trigger('click');
-
-        const dialog = wrapper.find('#faq-dialog .wikit-Dialog');
+        const dialog = wrapper.find('#faq-dialog.cdx-dialog');
         expect(dialog.isVisible()).toBe(true);
     });
 
     it('validates that textarea input is not empty', async () => {
         const itemsInput = '';
 
-        const store = new Vuex.Store({});
-
         const wrapper = mount(Home, {
-            mocks,
-            localVue,
-            store,
+            global: {
+                mocks,
+                plugins: [createTestingPinia()]
+            },
             data() {
                 return {
                     form: {
@@ -91,14 +101,13 @@ describe('Home.vue', () => {
     });
 
     it('validates that items in textarea input dont exceed the maximum number of ids', async () => {
-        const store = new Vuex.Store();
-
         const itemsInput = Array(MAX_NUM_IDS + 2).fill('Q21').join('\n');
 
         const wrapper = mount(Home, {
-            mocks,
-            localVue,
-            store,
+            global: {
+                mocks,
+                plugins: [createTestingPinia()],
+            },
             data() {
                 return {
                     form: {
@@ -116,14 +125,13 @@ describe('Home.vue', () => {
     });
 
     it('validates that items in textarea input are well-formed', async () => {
-        const store = new Vuex.Store();
-
         const itemsInput = 'L12345';
 
         const wrapper = mount(Home, {
-            mocks,
-            localVue,
-            store,
+            global: {
+                mocks,
+                plugins: [createTestingPinia()],
+            },
             data() {
                 return {
                     form: {
@@ -142,9 +150,10 @@ describe('Home.vue', () => {
 
     it('shows error message upon serverside validation errors', async () => {
         mocks.$page.props.errors = { 'someKey' : 'someError'}
-        const store = new Vuex.Store();
-        const wrapper = mount(Home, { mocks, localVue, store });
-
+        const wrapper = mount(Home, { global: {
+                mocks,
+                plugins: [createTestingPinia()]
+            }});
         const errorMessage = wrapper.find('#message-section .wikit-Message--error.wikit');
         expect(errorMessage.isVisible()).toBe(true);
     });
