@@ -1,5 +1,5 @@
 const mix = require('laravel-mix');
-
+const path = require('path');
 /*
  |--------------------------------------------------------------------------
  | Mix Asset Management
@@ -12,7 +12,35 @@ const mix = require('laravel-mix');
  */
 
 mix.ts('resources/js/app.ts', 'public/js')
-    .vue({ version: 2 })
+    .webpackConfig({
+        resolve: {
+            alias: {
+                vue: '@vue/compat',
+                '@vue/composition-api': '@vue/compat',
+                '@wmde/wikit-vue-components':
+                    '@wmde/wikit-vue-components/dist/wikit-vue-components-vue3compat.common.js',
+                'wikit-dist': path.resolve(__dirname, './node_modules/@wmde/wikit-vue-components/dist'),
+            }
+        },
+        module: {
+            rules: [
+                {
+                    test: /\.vue$/,
+                    loader: 'vue-loader',
+                    options: {
+                        compilerOptions: {
+                            compatConfig: {
+                                MODE: 3,
+                                COMPILER_V_ON_NATIVE: true,
+                                COMPILER_V_BIND_SYNC: true
+                            }
+                        }
+                    }
+                }
+            ]
+        }
+    })
+    .vue({ version: 3})
     .sass('resources/sass/app.scss', 'public/css')
     .sass('resources/sass/noscript.scss', 'public/css')
     .copyDirectory('resources/img', 'public/images')
